+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-## 1.0.0
+## 1.0.1
+- Added a example to package.
+- Added more documentation in atributes of package.
+- Added icon in `README.md`.
+- README.md now has a icon and header.
+- repository url added to yaml.
+- issue_tracker url added to yamk.
 
+## 1.0.0 
 - Initial version.